link to friendlier new-subscription page instead of "Contact us"
After https://github.com/sourcegraph/enterprise/pull/13708 is merged, the new-subscription page on Sourcegraph.com will be friendlier: you can price out a subscription even if you're not signed in. Previously it would dump you on a sign-in page, so we didn't feel comfortable linking directly to it from this pricing page.