transition from remote-first to all-remote (#347)
* transition from remote-first to all-remote

Proposes transitioning Sourcegraph from a remote-first company to an all-remote company. We briefly discussed this at company meeting on 2019-11-11 and then [on Slack in #general](https://sourcegraph.slack.com/archives/C02FSM7DU/p1573498881167500). This is the actual proposal to change this aspect of Sourcegraph (in line with [Changing or defining a process](https://about.sourcegraph.com/handbook/usage#changing-or-defining-a-process) in the handbook).

Remote work has been a part of Sourcegraph since very early on, and it has worked extremely well. More than half of our team is remote now, and we anticipate the proportion of remote teammates will grow because the population of talented people in SF is tiny compared to the population of talented people worldwide.

We transitioned from remote-friendly to remote-first earlier this year. That switch was successful. Several teammates who otherwise would have felt obligated to work from our SF office instead chose to move to a location they favor over SF or work from home despite being based in the SF Bay Area. Those people are happier and more productive because of it. The switch also helped us attract new talented remote candidates who prefer remote-first companies to remote-friendly companies (which are hit-or-miss, sometimes they claim to be remote-friendly but without the forcing function of remote-first let their communication practices privilege local folks).

Now that we are remote-first, why go all-remote?

1. Having an office causes many SF-based people to feel obligated to come into the office more than is optimal. They tell us they do this because they "don't want the office to feel empty" despite reporting that they would feel more productive and happier working from home more of those days. This is an admirable impulse, but we don't want people to feel guilty about not coming into the office and therefore choose a non-optimal arrangement.
   - Could we just *really* clearly communicate to these people that they aren't obligated to come into the office? We've tried that, but it hasn't changed their behavior (they still feel obligated, and people still feel bad that the office is empty).
   - Could we just get a smaller official office location? We could try that, and it'd probably make the situation better but not completely address it. For people who like having an o;5Dffice, we instead suggest they coordinate working out of a private office(s) in a coworking space with other like-minded teammates. That yields a very similar outcome to the people who prefer an office without affecting other teammates.
1. All-remote is more appealing to talented people worldwide than remote-first because it means they are on a level playing field as everyone else at the company (and not at a slight, possibly in perception only, disadvantage to local teammates). Being all-remote will help us continue to grow our team with world-class teammates.

What about people who prefer having an office with other teammates? Some SF-based teammates expressed this. See the "Could we just get a smaller office?" point above. You can coordinate with other SF-based teammates to get a private office(s) at a coworking space.

Notes:

- We're going to document a lot more of our practices and policies for remote work, such as how new teammate onboarding changes when everyone is remote. These will come in future edits to the handbook.
- Now that we're all remote, we don't need to use the term "distributed" in preference to "remote".

* finish sentence

* Add caveat for office (#404)

Language not yet finalized, waiting on legal feedback

* more clarification on office