Edit Campaigns page and refine Vision/Mission (#628)
After the discussions yesterday I think it makes sense to include what
we _don't_ do. See my comment here: https://github.com/sourcegraph/about/pull/623#issuecomment-593560930

I'm still not sure about the "Vision"/"Mission" naming and this still
does not include the things that might at some point come in the future
or anticipate things that might change the vision (like
linters/diagnostics/etc.), but I wanted to add this so that we have this
written down somewhere.

As always, we can still change it.