Add Working with design files + heading update (#1187)
* Add Working with design files + heading update 

I've added 'Working with design files' section that describes all the rules that we've agreed on recently when it comes to working with Figma files, naming conventions and Project Tools.

In addition, I've changed slightly the formatting of the 'Process segments' section - using bold text for the main steps in the process increases the readability of this document.

Please, share your feedback regarding both the content and the form :) Thank you!

* Update handbook/product/design/design-process.md

Co-authored-by: Christina Forney <christina@sourcegraph.com>

* Update handbook/product/design/design-process.md

Co-authored-by: Christina Forney <christina@sourcegraph.com>

* Update handbook/product/design/design-process.md

Co-authored-by: Christina Forney <christina@sourcegraph.com>

* Update handbook/product/design/design-process.md

Co-authored-by: Christina Forney <christina@sourcegraph.com>

* Update handbook/product/design/design-process.md

Co-authored-by: Christina Forney <christina@sourcegraph.com>

* Add Pages to 'File structure' according

Changes according to Rob's suggestion. @rrhyne please review. Thank you!

Co-authored-by: Rob Rhyne <rob@sourcegraph.com>
Co-authored-by: Christina Forney <christina@sourcegraph.com>