Make it clearer who to ping about JAMF (#4125)
* Make it clearer which Cecily to ping about JAMF

The text previously showed only Cecily without a surname or a slack handle. I changed the decsription to include a full name and a link to the profile to make it explicit, who to contact. As I was not sure when going through the onboarding guide.

* Updated instructions for setting up Jamf Now

New process is now handled by Tech Ops

Co-authored-by: Cecily Black <68344143+CecilyBlack@users.noreply.github.com>